delte unused pairs
